+++ libs/gesha-core/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+gesha-rust-types = { path = "../../libs/gesha-rust-types" }
 openapi-types = { path = "../../libs/openapi-types" }
 
 console = "0.15.7"
