+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/components_shape/transform_schemas/convert_one_of.rs
@@ -6,7 +6,7 @@ use crate::conversions::v3_0::to_rust_type::components::ComponentsShape;
 use crate::conversions::Error::ReferenceObjectNotFound;
 use crate::conversions::Result;
 use crate::misc::TryMap;
-use crate::targets::rust_type::{EnumVariantName, SerdeAttribute};
+use gesha_rust_types::{EnumVariantName, SerdeAttribute};
 use SerdeAttribute::Untagged;
 
 pub fn convert_one_of(mut shapes: ComponentsShape) -> Result<ComponentsShape> {
