+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/definition_shape.rs
@@ -4,7 +4,7 @@ use crate::conversions::v3_0::to_rust_type::components::schemas::{
     TypeShape,
 };
 use crate::conversions::Result;
-use crate::targets::rust_type::{Definition, NewTypeDef, StructDef, StructField};
+use gesha_rust_types::{Definition, NewTypeDef, StructDef, StructField};
 
 #[derive(Clone, Debug)]
 pub enum DefinitionShape {
