+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/mod_shape.rs
@@ -1,7 +1,7 @@
 use crate::conversions::v3_0::to_rust_type::components::schemas::{DefinitionShape, TypeShape};
 use crate::conversions::Result;
 use crate::misc::TryMap;
-use crate::targets::rust_type::{Definitions, ModDef, ModuleName, Package};
+use gesha_rust_types::{Definitions, ModDef, ModuleName, Package};
 use openapi_types::v3_0::ComponentName;
 
 #[derive(Clone, Debug)]
