+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/schemas_shape.rs
@@ -3,7 +3,7 @@ use crate::conversions::v3_0::to_rust_type::components::schemas::{
     OneOfShape, Ref, StructShape, TypeHeaderShape, TypeShape,
 };
 use crate::conversions::Result;
-use crate::targets::rust_type::ModDef;
+use gesha_rust_types::ModDef;
 use openapi_types::v3_0::{ComponentName, SchemaCase, SchemaObject, SchemasObject};
 use std::ops::Not;
