+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/type_header_shape.rs
@@ -1,4 +1,4 @@
-use crate::targets::rust_type::{DocComments, SerdeAttribute, TypeHeader};
+use gesha_rust_types::{DocComments, SerdeAttribute, TypeHeader};
 use heck::ToUpperCamelCase;
 use openapi_types::v3_0::{ComponentName, SchemaObject};
