+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/type_path.rs
@@ -1,4 +1,4 @@
-use crate::targets::rust_type::DataType;
+use gesha_rust_types::DataType;
 use std::fmt::{Display, Formatter};
 
 #[derive(Debug, Clone)]
