+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/type_shape.rs
@@ -3,7 +3,7 @@ use crate::conversions::v3_0::to_rust_type::components::schemas::TypeShape::{Inl
 use crate::conversions::v3_0::to_rust_type::components::schemas::{Optionality, Ref, TypePath};
 use crate::conversions::Error::UnknownFormat;
 use crate::conversions::Result;
-use crate::targets::rust_type::DataType;
+use gesha_rust_types::DataType;
 use openapi_types::v3_0::SchemaCase;
 use openapi_types::v3_0::SchemaCase::{Reference, Schema};
 use openapi_types::v3_0::{FormatModifier, OpenApiDataType, SchemaObject};
