+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/mod.rs
@@ -5,7 +5,7 @@ use crate::conversions::v3_0::to_rust_type::components::request_bodies::RequestB
 use crate::conversions::v3_0::to_rust_type::components::schemas::SchemasShape;
 use crate::conversions::v3_0::to_rust_type::components::ComponentsShape;
 use crate::conversions::{Result, ToRustType};
-use crate::targets::rust_type::{DataType, Modules};
+use gesha_rust_types::{DataType, Modules};
 use openapi_types::v3_0::{ComponentsObject, Document};
 
 impl ToRustType<Document> for Modules {
