+++ libs/gesha-core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod gateway;
-pub mod targets;
 
 mod conversions;
 mod misc;
