+++ libs/gesha-core/src/renderer/rust_type/render_media_type.rs
@@ -1,6 +1,6 @@
 use crate::render;
 use crate::renderer::Result;
-use crate::targets::rust_type::MediaTypeDef;
+use gesha_rust_types::MediaTypeDef;
 use std::io::Write;
 
 pub fn render_media_type<W: Write>(mut write: W, x: MediaTypeDef) -> Result<()> {
