--- libs/gesha-core/src/targets/mod.rs
@@ -1,1 +1,0 @@
-pub mod rust_type;
