+++ libs/gesha-rust-types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "gesha-rust-types"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+heck = "0.4.1"
+indexmap = "2.0.0"
