+++ libs/gesha-rust-types/src/error_def.rs
